fix(gemini): reorganize code that works with columns
