+++ routing_key_test.go
@@ -474,7 +474,7 @@ func TestRoutingKey(t *testing.T) {
 func createPkColumns(cnt int, prefix string) Columns {
 	var cols Columns
 	for i := 0; i < cnt; i++ {
-		cols = append(cols, ColumnDef{
+		cols = append(cols, &ColumnDef{
 			Name: genColumnName(prefix, i),
 			Type: TYPE_INT,
 		})
